refactor: split mission-control into sections and a pure model module
MissionControl.tsx was 701 lines mixing GitHub data loading, view-model
assembly, formatters and five page sections in one file.

Pure logic moves to lib/mission-control/model.ts, joining the eight tested
modules already there: RepoData/BenchmarkGlobal, fmtInt/round1/pctOrDash,
safe, loadAll, buildModel and Model. actionById moves beside WRITE_ACTIONS
in write-actions.ts.

Markup splits into Stat (15 call sites), Section (5), ExternalLink,
MissionControlSkeleton and five section components. MissionControl.tsx keeps
state, effects, confirm/runAction and composition at 235 lines.

Behavior-preserving: rendered DOM, text, class names and interaction flow
are unchanged. The six existing end-to-end tests pass untouched.
